fixed restful calls in dikb evidence

+++ src/main/java/org/ohdsi/webapi/DIKB/SourceDBModel.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class SourceDBModel {
 
-		  @JsonProperty("name")
-		  public String sourceType;
-		  
-		  @JsonProperty("fullname")
-		  public String fullname;
-		  
-		  @JsonProperty("value")
-		  public int sourceNum;
+    @JsonProperty("name")
+    public String sourceType;
+    
+    @JsonProperty("fullname")
+    public String fullname;
+    
+    @JsonProperty("value")
+    public Long sourceNum;
 		  
 }
